+++ projects/rocthrust/CHANGELOG.md
@@ -7,6 +7,8 @@ Documentation for rocThrust available at
 
 ### Added
 
+* Add `thrust::reduce_into`.
+
 ### Changed
 
 * Defined `THRUST_COMP_DEPR_IMPL` with `_CCCL_WARNING`.